alot of changes in this one.
completed update #1

+++ Source/src/com/client/common/updateViewer.java
@@ -6,12 +6,69 @@ import java.awt.*;
 import java.awt.event.*;
 import java.io.*;
 
-public class updateViewer extends JFrame
+public class updateViewer extends JFrame implements Runnable
 {
 	private static final long serialVersionUID = 1L;
 	private JButton okButton = null;
 	private File update = null;
+	JTextArea updateText = null;
+	Thread uThread;
 	
+	///---ETL Constructor---///
+	public updateViewer()
+	{
+		//Threading
+		uThread = new Thread(this);
+		uThread.setName("Update Viewer");
+	}
+	
+	///---Running---///
+	public void run()
+	{
+		//Frame
+		this.setVisible(false);
+		this.setLocation(100, 100);
+		this.setSize(480,640);
+		this.setBackground(Color.CYAN);
+		this.setLayout(null);
+		this.addKeyListener(new keyListener());
+		this.setTitle("About This Update");
+		
+		//Text Area
+		updateText = new JTextArea();
+		updateText.setEditable(false);
+		updateText.setSize(440, 540);
+		updateText.setLocation(10, 10);
+		updateText.addKeyListener(new keyListener());
+		updateText.setAutoscrolls(true);
+		this.add(updateText);
+		
+		//Okay Button
+		okButton = new JButton();
+		okButton.setText("OK");
+		okButton.addActionListener(new actionListener());
+		okButton.setLocation(10, 555);
+		okButton.setSize(440, 40);
+		this.add(okButton);
+		okButton.setEnabled(false);
+		
+		this.setVisible(true);
+		
+		try
+		{
+			uThread.sleep(500);
+		}
+		catch(Exception e)
+		{
+			//Ignore
+		}
+	}
+	public void start()
+	{
+		uThread.start();
+	}
+	
+	///---DBM Constructor---///
 	public updateViewer(File file)
 	{
 		this.setVisible(false);
@@ -27,7 +84,7 @@ public class updateViewer extends JFrame
 		update = file;
 		
 		//Text Area
-		JTextArea updateText = new JTextArea();
+		updateText = new JTextArea();
 		updateText.setEditable(false);
 		updateText.setSize(440, 540);
 		updateText.setLocation(10, 10);
@@ -89,18 +146,34 @@ public class updateViewer extends JFrame
 		this.add(okButton);
 	}
 	
+	///---ETL Methods---///
+	public void etlComplete()
+	{
+		updateText.setText("Routine Update Complete\nThank You!");
+		okButton.setEnabled(true);
+	}
+	public void etlUpdate(String info)
+	{
+		updateText.setText(updateText.getText() + "\n" + info);
+	}
+	
+	///---Action Listener---///
 	private class actionListener implements ActionListener
 	{
 		public void actionPerformed(ActionEvent ae)
 		{
 			if(ae.getSource() == okButton)
 			{
-				update.delete();
+				if(update != null)
+				{
+					update.delete();
+				}
 				dispose();
 			}
 		}
 	}
 	
+	///---Key Listener---///
 	private class keyListener implements KeyListener
 	{
 		public void keyPressed(KeyEvent ke)
